stdout: avoid body content to be truncated if first byte is null
ap_scan_script_header_err_strs uses getsfunc_STRING in order to find
lines in buffer pointed by data. However this function does not expect
to have binary data in the passed arguments but null terminated
strings.

In some cases, the fastcgi server (php-fpm) groups http headers and
the begining of the body content in the same packet. In case the body
begins with a null byte (like mov files) the
getsfunc_STRING function skips all the rest of the data
after parsing the headers. This makes the returned body truncated.

The exact line causing this bug is located at
httpd-2.4.10/server/util_script.c:759

    759 if (!strs->curpos[t]) {
    760     ++strs->arg;
    761     strs->curpos = va_arg(strs->args, const char *);
    762 }

The rest of the data is skipped because just after the \n there is a
null byte (the first byte of the content but not the end of the data
to parse).

We fix this behavior by implementing a new getsfunc (getsfunc_MEMORY)
that uses the payload_len as the limit of the data to parse.